Stream -> StreamCore (Shard args) + Stream (all).
